backend/requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
pypdf==5.4.0
requests==2.32.3
python-dotenv==1.0.1 